requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
paddlepaddle==2.5.2
paddleocr==2.7.0.3
python-multipart==0.0.6
pillow==10.1.0
numpy==1.24.3
opencv-python==4.6.0.66
pydantic-settings==2.0.3
aiohttp==3.9.1
PyMuPDF==1.23.8
PyPDF2==3.0.1
python-docx==1.1.0
